update approve product

+++ app/src/main/java/com/example/greenbuyapp/ui/profile/editProfile/infomation/CustomerInformationActivity.kt
@@ -204,7 +204,7 @@ class CustomerInformationActivity : BaseActivity<ActivityCustomerInformationBind
             avatarUrl, // Có thể null nếu không chọn ảnh mới
             ho,
             ten,
-            phone.toInt(),
+            phone,
             selectedBirthDate!!
         )
